+++ app/src/main/java/com/gtappdevelopers/findtoday/Dao.java
@@ -5,10 +5,8 @@ import androidx.room.Delete;
 import androidx.room.Insert;
 import androidx.room.Query;
 import androidx.room.Update;
-
 import java.util.List;
 
-//Adding annotation to our Dao class
 @androidx.room.Dao
 public interface Dao {
     @Insert
