+++ app/src/main/java/com/gtappdevelopers/findtoday/FinDatabase.java
@@ -14,7 +14,6 @@ public abstract class FinDatabase extends RoomDatabase {
     private static FinDatabase instance;
     public abstract Dao Dao();
 
-
     //on below line we are getting instance for our database.
     public static synchronized FinDatabase getInstance(Context context) {
         //below line is to check if the instance is null or not.
@@ -40,14 +39,11 @@ public abstract class FinDatabase extends RoomDatabase {
 
     //we are creating an async task class to perform task in background.
     private static class PopulateDbAsyncTask extends AsyncTask<Void, Void, Void> {
-
         PopulateDbAsyncTask(FinDatabase instance) {
             Dao dao = instance.Dao();
         }
-
         @Override
         protected Void doInBackground(Void... voids) {
-
             return null;
         }
     }
