+++ app/src/main/java/com/gtappdevelopers/findtoday/MainActivity.java
@@ -3,16 +3,12 @@ package com.gtappdevelopers.findtoday;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.app.ActivityCompat;
-import androidx.core.content.ContextCompat;
 import androidx.lifecycle.Observer;
-import androidx.lifecycle.ViewModelProviders;
+import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.ItemTouchHelper;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
-import android.Manifest;
 import android.content.Intent;
-import android.content.pm.PackageManager;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Toast;
@@ -20,12 +16,10 @@ import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import java.util.List;
 
 public class MainActivity extends AppCompatActivity {
-    //creating a variables for our recycler view.
     private ViewModal viewmodal;
     private static final int ADD_DESP_REQUEST = 1;
     public static final int EDIT_DESP_REQUEST = 2;
-    //private static final int EDIT_DESP_REQUEST = 2;
-    private static final int SEARCH_DESP_REQUEST = 3; // Define um requestCode para BuscarFinActivity
+    private static final int SEARCH_DESP_REQUEST = 3;
     private static final int REQUEST_CODE_WRITE_EXTERNAL_STORAGE = 1001;
 
     @Override
@@ -33,19 +27,8 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        if (ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED) {
-            ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE}, REQUEST_CODE_WRITE_EXTERNAL_STORAGE);
-        } else {
-            boolean backupSuccess = DatabaseBackup.backupDatabase(this);
-            if (backupSuccess) {
-                Toast.makeText(this, "Backup do banco de dados concluído com sucesso.", Toast.LENGTH_SHORT).show();
-            } else {
-                Toast.makeText(this, "Falha ao realizar o backup do banco de dados.", Toast.LENGTH_SHORT).show();
-            }
-        }
-
         FloatingActionButton fab = findViewById(R.id.idFABAdd);
-        FloatingActionButton fab2 = findViewById(R.id.idFABAdd2);
+
         fab.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -53,6 +36,8 @@ public class MainActivity extends AppCompatActivity {
                 startActivityForResult(intent, ADD_DESP_REQUEST);
             }
         });
+
+        FloatingActionButton fab2 = findViewById(R.id.idFABAdd2);
         fab2.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -66,7 +51,7 @@ public class MainActivity extends AppCompatActivity {
         FinRV.setHasFixedSize(true);
         final FinRVAdapter adapter = new FinRVAdapter();
         FinRV.setAdapter(adapter);
-        viewmodal = ViewModelProviders.of(this).get(ViewModal.class);
+        viewmodal = new ViewModelProvider(this).get(ViewModal.class);
 
         viewmodal.getallDesp().observe(this, new Observer<List<FinModal>>() {
             @Override
@@ -104,7 +89,6 @@ public class MainActivity extends AppCompatActivity {
 
     }
 
-
     @Override
     protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
@@ -131,6 +115,7 @@ public class MainActivity extends AppCompatActivity {
             String fontDesp = data.getStringExtra(NewFinActivity.EXTRA_FONT_DESP);
             String despDescr = data.getStringExtra(NewFinActivity.EXTRA_DESCR_DESP);
             String dataDesp = data.getStringExtra(NewFinActivity.EXTRA_DURATION);
+
             FinModal model = new FinModal(valorDesp, tipoDesp, fontDesp, despDescr, dataDesp);
             model.setId(id);
             viewmodal.update(model);
@@ -144,20 +129,5 @@ public class MainActivity extends AppCompatActivity {
 
     }
 
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        if (requestCode == REQUEST_CODE_WRITE_EXTERNAL_STORAGE) {
-            if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                boolean backupSuccess = DatabaseBackup.backupDatabase(this);
-                if (backupSuccess) {
-                    Toast.makeText(this, "Backup do banco de dados concluído com sucesso.", Toast.LENGTH_SHORT).show();
-                } else {
-                    Toast.makeText(this, "Falha ao realizar o backup do banco de dados.", Toast.LENGTH_SHORT).show();
-                }
-            }   else {
-                Toast.makeText(this, "Permissão de escrita no armazenamento externo negada.", Toast.LENGTH_SHORT).show();
-            }
-        }
-    }
+
 }
