+++ app/src/main/java/com/gtappdevelopers/findtoday/ResumoDespActivity.java
@@ -22,4 +22,5 @@ public class ResumoDespActivity extends AppCompatActivity {
             }
         });
     }
+
 }
