+++ app/src/main/res/layout/activity_new_fin.xml
@@ -77,5 +77,14 @@
         android:text="Consultar Resumo"
         android:textAllCaps="false" />
 
+    <Button
+        android:id="@+id/idBtnbkp"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_margin="10dp"
+        android:padding="5dp"
+        android:text="Realizar bkp"
+        android:textAllCaps="false" />
+
 
 </LinearLayout>
